made changes to chatbot.js(works properly) created a README.md specified .env to be hidden in .gitignore
